refactor[#242] : constants 폴더 오타 수정
